Normalize scores to the best score for each game
Currently, scores are calculated as the distance between the actual game
score and the player's prediction. The best score per game is the one
closest to the actual game score (i.e., the one with the lowest
distance).

This change takes the best score for each game and normalizes all the
other scores to it. In other words, now the scores are calculated as teh
distance between the best score and the player's prediciton.